+++ packages/dsw-tdk/tests/test_cmd_put.py
@@ -11,6 +11,7 @@ def test_put_ok(fixtures_path: pathlib.Path, dsw_env: dict):
     runner = click.testing.CliRunner()
     template_path = fixtures_path / 'test_example01'
     result = runner.invoke(main, args=['put', template_path.as_posix()], env=dsw_env)
+    print(result.stdout)
     assert result.exit_code == 0
 
 
@@ -35,7 +36,7 @@ def test_put_bad_token(fixtures_path: pathlib.Path, dsw_api_url: str):
     runner = click.testing.CliRunner()
     template_path = fixtures_path / 'test_example01'
     env_vars = {
-        'DSW_API': dsw_api_url,
+        'DSW_API_URL': dsw_api_url,
         'DSW_API_KEY': 'foo',
     }
     result = runner.invoke(main, args=['put', template_path.as_posix()], env=env_vars)
@@ -47,7 +48,7 @@ def test_put_bad_url(fixtures_path: pathlib.Path, dsw_api_key: str):
     runner = click.testing.CliRunner()
     template_path = fixtures_path / 'test_example01'
     env_vars = {
-        'DSW_API': 'http://localhost:33333',
+        'DSW_API_URL': 'http://localhost:33333',
         'DSW_API_KEY': dsw_api_key,
     }
     result = runner.invoke(main, args=['put', template_path.as_posix()], env=env_vars)