+++ packages/dsw-database/CHANGELOG.md
@@ -8,6 +8,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 
+## [3.27.0]
+
+Released for version consistency with other DSW tools.
+
 ## [3.26.1]
 
 Released for version consistency with other DSW tools.
@@ -147,3 +151,4 @@ Released for version consistency with other DSW tools.
 [3.25.1]: /../../tree/v3.25.1
 [3.26.0]: /../../tree/v3.26.0
 [3.26.1]: /../../tree/v3.26.1
+[3.27.0]: /../../tree/v3.27.0