+++ packages/dsw-tdk/CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [3.27.0]
+
+### Changed
+
+- Removed legacy support for authentication using credentials
+
 ## [3.26.1]
 
 Released for version consistency with other DSW tools.
@@ -340,3 +346,4 @@ Initial DSW Template Development Kit (versioned as part of the [DSW platform](ht
 [3.25.0]: /../../tree/v3.25.0
 [3.26.0]: /../../tree/v3.26.0
 [3.26.1]: /../../tree/v3.26.1
+[3.27.0]: /../../tree/v3.27.0